requirements.txt
opencv_python==4.1.1.26
torch==1.3.0
facenet_pytorch==0.3.1
numpy==1.17.2
torchvision==0.4.1
Pillow==6.2.0
celluloid==0.2.0
matplotlib==3.1.1
tqdm==4.36.1
